Major changes to plot function: grid and toggle graph functions
Plot function is not not focused on any specific plot, but on all plots. Calls records_to_dict function (minor changes there too: removed hardcoded fox
and cat labels) and for every label creates a plot, then saves to self.plots or whatever. Dates are now rotated. The line below fixes the bug: the
dates are cut off by the end of the window. Remove the line and show the bug